updated demo , test-app deleted, only missing is team photos
