+++ quell-client/quell-client/src/helpers/joinResponses.js
@@ -1,12 +1,21 @@
+///////////////////////////////////////////////////////////////////////////////////
+////////THIS CODE IS DEPRICATED SINCE THE INTRODUCTION OF LOKIJS///////////////////
+//////////////////////////////////////////////////////////////////////////////////
+
 /**
- joinResponses combines two objects containing results from the cached response and fetched (uncached) and outputs a single array response that will ultimately be formatted and delivered to the client. 
+ joinResponses combines two objects containing results from the cached response and fetched (uncached) and outputs a single array response that will ultimately be formatted and delivered to the client.
  the copied Proto parameter sets a reference to combine the fields in the same order as the original query.
- @param {object} cacheResponse - JavaScript object containing the result from the cached response 
- @param {object} serverResponse - JavaScript object containing the fetched (uncached) result 
+ @param {object} cacheResponse - JavaScript object containing the result from the cached response
+ @param {object} serverResponse - JavaScript object containing the fetched (uncached) result
  @param {object} queryProto - Prototype object from query
  */
 
-function joinResponses(cacheResponse, serverResponse, queryProto, fromArray = false) {
+function joinResponses(
+  cacheResponse,
+  serverResponse,
+  queryProto,
+  fromArray = false
+) {
   // initialize a "merged response" to be returned
   let mergedResponse = {};
   // loop through fields object keys, the "source of truth" for structure
@@ -14,18 +23,22 @@ function joinResponses(cacheResponse, serverResponse, queryProto, fromArray = fa
   // first loop for different queries on response
   for (const key in queryProto) {
     // for each key, check whether data stored at that key is an array or an object
-    const checkResponse = cacheResponse.hasOwnProperty(key) ? cacheResponse : serverResponse;
+    const checkResponse = cacheResponse.hasOwnProperty(key)
+      ? cacheResponse
+      : serverResponse;
 
     if (Array.isArray(checkResponse[key])) {
       // merging data stored as array
 
-      if (cacheResponse.hasOwnProperty(key) && serverResponse.hasOwnProperty(key)) {
+      if (
+        cacheResponse.hasOwnProperty(key) &&
+        serverResponse.hasOwnProperty(key)
+      ) {
         // if # of keys is not the same, cache was missing data for each object, need to merge cache objects with server objects
-        
+
         // iterate over an array
         const mergedArray = [];
         for (let i = 0; i < checkResponse[key].length; i++) {
-
           // for each index of array, combine cache and server response objects
           const joinedResponse = joinResponses(
             { [key]: cacheResponse[key][i] },
@@ -39,15 +52,12 @@ function joinResponses(cacheResponse, serverResponse, queryProto, fromArray = fa
         }
         // set merged array to mergedResponse at key
         mergedResponse[key] = mergedArray;
-      }
-      else if (cacheResponse.hasOwnProperty(key)) {
+      } else if (cacheResponse.hasOwnProperty(key)) {
         mergedResponse[key] = cacheResponse[key];
-      }
-      else {
+      } else {
         mergedResponse[key] = serverResponse[key];
       }
-    }
-    else {
+    } else {
       // if not an array, it is a regular object data structure
 
       // object spread
@@ -58,22 +68,24 @@ function joinResponses(cacheResponse, serverResponse, queryProto, fromArray = fa
       } else {
         // if the object comes from an array, we do not want to assign to a key as per GQL spec
         // results in [{fields}, {fields}, {fields}]
-        mergedResponse = { ...cacheResponse[key], ...serverResponse[key] }
+        mergedResponse = { ...cacheResponse[key], ...serverResponse[key] };
       }
-      
+
       // loop through fields on queryProto
       for (const fieldName in queryProto[key]) {
-
         // check for nested objects
-        if (typeof queryProto[key][fieldName] === 'object' && !fieldName.includes('__')) {
+        if (
+          typeof queryProto[key][fieldName] === 'object' &&
+          !fieldName.includes('__')
+        ) {
           // recurse joinResponses on that object to create deep copy on mergedResponse
 
           const mergedRecursion = joinResponses(
             { [fieldName]: cacheResponse[key][fieldName] },
-            { [fieldName]: serverResponse[key][fieldName] }, 
+            { [fieldName]: serverResponse[key][fieldName] },
             { [fieldName]: queryProto[key][fieldName] }
           );
-  
+
           // place on merged response
           mergedResponse[key] = { ...mergedResponse[key], ...mergedRecursion };
         }
