+++ quell-client/quell-client/src/helpers/mapGenerator.js
@@ -1,8 +1,10 @@
+/// this is the code used to take in the clients endpoint and generate the mutationMap,Map and queryTypeMap
+// it takes advantage of GraphQl's built in introspection.
+
 const mapGenerator = async (endpoint) => {
   //for queryTypeMap
 
-  console.log('in map generator');
-
+  console.log('here in mapGenerator');
   const mapGeneratorHelper = async (endpoint, query, type) => {
     const obj = {};
 
@@ -15,30 +17,40 @@ const mapGenerator = async (endpoint) => {
     };
 
     const serverResponse = await fetch(endpoint, fetchOptions);
+    console.log('line 20');
+    console.log(serverResponse);
     const parsedData = await serverResponse.json();
+    console.log('line 22', parsedData);
 
     let fieldsArray;
     let typesList;
     if (type === 'queryType') {
       fieldsArray = parsedData.data.__schema.queryType.fields;
+      console.log('line 26:', fieldsArray);
     } else if (type === 'mutationType') {
       fieldsArray = parsedData.data.__schema.mutationType.fields;
+      console.log(fieldsArray);
     } else {
       typesList = parsedData.data.__schema.types;
     }
 
-    if (type === 'queryType' || type === 'mutationType') {
+    console.log('line 32');
+    if (type === 'queryType') {
+      for (let types of fieldsArray) {
+        let queryType = types.name;
+        let queryTypeValue = types.type.ofType.name;
+        obj[queryType] = queryTypeValue;
+      }
+
+      return obj;
+    }
+
+    if (type === 'mutationType') {
       for (let types of fieldsArray) {
         let queryType = types.name;
         let queryTypeValue = types.type.name;
-        if (queryTypeValue) {
-          obj[queryType] = queryTypeValue;
-        } else {
-          queryTypeValue = types.type.ofType.ofType.ofType.name;
-          obj[queryType] = queryTypeValue;
-        }
+        obj[queryType] = queryTypeValue;
       }
-
       return obj;
     }
 
@@ -50,6 +62,8 @@ const mapGenerator = async (endpoint) => {
       'ID',
       'Query',
       'Mutation',
+      'RootQueryType',
+      'RootMutationType',
       '__Type',
       '__Field',
       '__EnumValue',
@@ -65,6 +79,7 @@ const mapGenerator = async (endpoint) => {
       (type) => !builtInTypes.includes(type.name)
     );
 
+    console.log('line 74:', customTypes);
     for (let types of customTypes) {
       let queryType = types.name;
       let queryTypeValue = types.name;
@@ -74,19 +89,35 @@ const mapGenerator = async (endpoint) => {
     return obj;
   };
 
+  //   const queryForQueryType = `{
+  // __schema{
+  //   queryType{
+  //     fields{
+  //         name
+  //         type{
+  //           name
+  //           ofType{
+  //             ofType{
+  //               ofType{
+  //                 name
+  //               }
+  //             }
+  //           }
+  //         }
+  //       }
+  //     }
+  //   }
+  // }`;
+
   const queryForQueryType = `{
 __schema{
   queryType{
     fields{
         name
-        type{
+        type {
           name
           ofType{
-            ofType{
-              ofType{
-                name
-              }
-            }
+            name
           }
         }
       }
@@ -132,6 +163,10 @@ __schema{
   );
 
   const map = await mapGeneratorHelper(endpoint, queryforMap, 'map');
+
+  console.log(map);
+  console.log(mutationMap);
+  console.log(queryTypeMap);
 
   return { map, mutationMap, queryTypeMap };
 };
