--- test-app/client/message-board/src/App.css
@@ -1,109 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: Arial, sans-serif;
-}
-#clear-btn-container {
-  display: flex;
-  /* width: 80%; */
-  justify-content: space-around;
-}
-#container {
-  position: relative;
-  margin: 2em auto;
-  width: 800px;
-  max-width: 80%;
-}
-
-.cacheBoard {
-  width: 800px;
-  height: 500px;
-}
-
-#fetch {
-  background-color: #4c50af;
-  max-width: 20%;
-}
-#create {
-  background-color: #4c50af;
-  max-width: 20%;
-}
-#delete {
-  background-color: #4c50af;
-  max-width: 20%;
-}
-
-#update {
-  background-color: #4c50af;
-  max-width: 300px;
-}
-
-#clear {
-  background-color: #4c50af;
-  max-width: 20%;
-  align-self: center;
-}
-
-.updateInput {
-  max-width: 40%;
-}
-
-#cacheButton {
-  display: block;
-}
-
-.row input {
-  width: 80%;
-}
-
-button {
-  background-color: #4caf50; /* Green */
-  border: none;
-  color: white;
-  padding: 0.5em 1.2em;
-  text-align: center;
-  text-decoration: none;
-  /* display: inline-block; */
-  font-size: 16px;
-  float: right;
-  width: 20%;
-}
-
-input {
-  padding: 8px;
-}
-
-#desc {
-  list-style: none;
-  border: 1px solid #dfdfdf;
-  width: 100%;
-  min-height: 4em;
-  clear: both;
-  font-size: 15px;
-  line-height: 1.5em !important;
-  margin: 5px auto;
-  padding: 25px 20px 40px;
-}
-
-ul {
-  position: relative;
-  margin-top: 2em;
-}
-
-li {
-  list-style: none;
-  border: 1px solid #dfdfdf;
-  width: 100%;
-  min-height: 4em;
-  clear: both;
-  font-size: 20px;
-  line-height: 1.5em !important;
-  margin: 5px auto;
-  padding: 25px 20px 40px;
-}
-
-li span {
-  margin-top: 0.5em;
-  width: 80%;
-}
