--- test-app/client/message-board/src/quell-client/.gitignore
@@ -1,2 +1,0 @@
-node_modules
-/src/sandbox.js
