--- test-app/client/message-board/src/quell-client/__test__/Quellify.test.js
@@ -1,8 +1,0 @@
-const Quellify = require('../src/Quellify');
-
-
-xdescribe('Quellify.test.js', () => {
-  test.skip('Quellify.test.js', () => {
-    expect(Quellify()).toEqual({});
-  });
-});
