--- test-app/client/message-board/src/quell-client/package.json
@@ -1,123 +1,0 @@
-{
-  "name": "@quell/client",
-  "version": "3.1.1",
-  "description": "Quell is an open-source NPM package providing a light-weight caching layer implementation and cache invalidation for GraphQL responses on both the client- and server-side. Use Quell to prevent redundant client-side API requests and to minimize costly server-side response latency.",
-  "main": "src/Quellify.js",
-  "scripts": {
-    "test": "jest --forceExit"
-  },
-  "engines": {
-    "node": "14.17.5",
-    "npm": "8.1.0"
-  },
-  "author": "Quell",
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/open-source-labs/Quell/tree/master/quell-client"
-  },
-  "bugs": {
-    "url": "https://github.com/open-source-labs/Quell/issues"
-  },
-  "keywords": [
-    "cache-invalidation",
-    "LokiJS",
-    "graphQL",
-    "cache",
-    "caching",
-    "redis",
-    "batching",
-    "client-side",
-    "server-side",
-    "apollo",
-    "quell"
-  ],
-  "contributors": [
-    {
-      "name": "Chang Cai",
-      "url": "https://github.com/ccai89"
-    },
-    {
-      "name": "Robert Howton",
-      "url": "https://github.com/roberthowton"
-    },
-    {
-      "name": "Joshua Jordan",
-      "url": "https://github.com/jjordan-90"
-    },
-    {
-      "name": "Jinhee Choi",
-      "url": "https://github.com/jcroadmovie"
-    },
-    {
-      "name": "Nayan Parmar",
-      "url": "https://github.com/nparmar1"
-    },
-    {
-      "name": "Tashrif Sanil",
-      "url": "https://github.com/tashrifsanil"
-    },
-    {
-      "name": "Tim Frenzel",
-      "url": "(https://github.com/TimFrenzel"
-    },
-    {
-      "name": "Thomas Reeder",
-      "url": "https://github.com/nomtomnom"
-    },
-    {
-      "name": "Ken Litton",
-      "url": "https://github.com/kenlitton"
-    },
-    {
-      "name": "Robleh Farah",
-      "url": "https://github.com/farahrobleh"
-    },
-    {
-      "name": "Angela Franco",
-      "url": "https://github.com/ajfranco18"
-    },
-    {
-      "name": "Andrei Cabrera",
-      "url": "https://github.com/Andreicabrerao"
-    },
-    {
-      "name": "Dasha Kondratenko",
-      "url": "https://github.com/dasha-k"
-    },
-    {
-      "name": "Derek Sirola",
-      "url": "https://github.com/dsirola1"
-    },
-    {
-      "name": "Xiao Yu Omeara",
-      "url": "https://github.com/xyomeara"
-    },
-    {
-      "name": "Mike Lauri",
-      "url": "https://github.com/MichaelLauri"
-    },
-    {
-      "name": "Rob Nobile",
-      "url": "https://github.com/RobNobile"
-    },
-    {
-      "name": "Justin Jaeger",
-      "url": "https://github.com/justinjaeger"
-    },
-    {
-      "name": "Nick Kruckenberg",
-      "url": "https://github.com/kruckenberg"
-    }
-  ],
-  "homepage": "http://quellql.io",
-  "dependencies": {
-    "fs": "^0.0.1-security",
-    "graphql": "^15.3.0",
-    "jest": "^26.6.3",
-    "lokijs": "^1.5.12"
-  },
-  "devDependencies": {
-    "@types/jest": "^26.0.20"
-  }
-}
