--- test-app/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "app-assessment-mod-1",
-  "version": "1.0.0",
-  "main": "server.js",
-  "scripts": {
-    "merge": "node util/merge.js",
-    "start": "nodemon server/server-quell"
-  },
-  "author": "Brandon Mizuno",
-  "license": "ISC",
-  "devDependencies": {
-    "eslint": "^7.17.0",
-    "nodemon": "^2.0.15"
-  },
-  "dependencies": {
-    "@quell/client": "^3.1.1",
-    "@quell/server": "^2.3.1",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "express": "^4.17.2",
-    "express-graphql": "^0.12.0",
-    "graphql": "^15.8.0",
-    "mongoose": "^6.1.5",
-    "pg": "^8.7.1",
-    "redis": "^4.0.2",
-    "redis-server": "^1.2.2"
-  },
-  "keywords": [],
-  "description": ""
-}
