--- test-app/views/index.html
@@ -1,31 +1,0 @@
-<html>
-<head>
-  <meta charset="UTF-8">
-  <title>Message Board</title>
-  <link rel="stylesheet" href="css/style.css">
-  <script src="http://cdnjs.cloudflare.com/ajax/libs/jquery/3.0.0-beta1/jquery.min.js"></script>
-  <script type="module" src="./js/index.js"></script>
-
-</head>
-<body>
-  <div id="container">
-  <h1>Get StarWars Character</h1><h1 id="time"><h1>
-  <div id="new-message">
-    <div class="row">
-      <input id="id" placeholder="id"></input>
-      <button id="fetch">Fetch</button>
-      <input id="createName" placeholder="name"></input>
-      <button id="create">Create</button>
-      <input id="deleteid" placeholder="delete"></input>
-      <button id="delete">delete</button>
-    <div>
-    </div>
-    <ul id="character-list"></ul>
-    <div id='clear-btn-container'>
-      <button id='clear'>Clear Board</button>
-    </div>
-  </div>
-
-  </div>
-</body>
-</html>
